[모의 SW 역량테스트] 활주로 건설

+++ [모의 SW 역량테스트] 활주로 건설/main.cpp
@@ -58,7 +58,7 @@ int main(int argc, const char * argv[]) {
     ios_base :: sync_with_stdio(false);
     cin.tie(NULL);
     cout.tie(NULL);
-    short T;
+    int T;
     cin >> T;
     
     for (int z = 1; z <= T; ++z) {
